requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
fancycompleter==0.9.1
gunicorn==20.0.4
idna==2.10
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pdbpp==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
pyrepl==0.9.0
pytz==2020.1
requests==2.24.0
sentry-sdk==0.17.3
sqlparse==0.3.1
traitlets==5.0.3
urllib3==1.25.10
wcwidth==0.2.5
whitenoise==5.2.0
wmctrl==0.3
